requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.1 